feat: implement user avatar management with update and delete functionality
